add condition for delete selenoid container

+++ selenoid/update.sh
@@ -8,4 +8,6 @@ cd ${BASEDIR}
 echo downloading latest chrome/firefox/opera browser images
 ${BASEDIR}/bin/bash selenoid update --vnc --config-dir "${BASEDIR}" $*
 
+if [ docker inspect selenoid | grep "Running" ]; then
 docker rm -f selenoid
+fi
